extracted implementations of ISerializerHandler to separate files
