+++ src/LionWeb-CSharp/core/M1/ISerializerHandler.cs
@@ -61,40 +61,3 @@ public interface ISerializerHandler
     /// or <c>null</c> to skip serializing this property.</returns>
     string? UnknownDatatype(IReadableNode node, Feature property, object? value);
 }
-
-public class SerializerExceptionHandler : ISerializerHandler
-{
-    /// <inheritdoc />
-    public void DuplicateNodeId(IReadableNode n) =>
-        throw new ArgumentException($"nodes have same id '{n.GetId()}': {n}");
-
-    /// <inheritdoc />
-    public virtual Language? DuplicateUsedLanguage(Language a, Language b) =>
-        throw new ArgumentException(
-            $"different languages with same key '{a?.Key ?? b?.Key}' / version '{a?.Version ?? b?.Version}': {a}, {b}");
-
-    /// <inheritdoc />
-    public string? UnknownDatatype(IReadableNode node, Feature property, object? value) =>
-        throw new ArgumentException($"unsupported property: {property}", nameof(property));
-}
-
-public class SerializerIgnoringHandler : ISerializerHandler
-{
-    /// <inheritdoc />
-    public void DuplicateNodeId(IReadableNode n) =>
-        Console.WriteLine($"nodes have same id '{n.GetId()}': {n}");
-
-    /// <inheritdoc />
-    public virtual Language? DuplicateUsedLanguage(Language a, Language b)
-    {
-        Console.WriteLine(
-            $"different languages with same key '{a?.Key ?? b?.Key}' / version '{a?.Version ?? b?.Version}': {a}, {b}");
-        return a;
-    }
-
-    public string? UnknownDatatype(IReadableNode node, Feature property, object? value)
-    {
-        Console.WriteLine($"unsupported property: {property}", nameof(property));
-        return null;
-    }
-}
